Use the com.google.enterprise.adaptor.testing package.
* Replace the local copies of RecordingDocIdPusher and
  RecordingResponse with the library's.
* Replace WrapperAdaptorContext with UnsupportedAdaptorContext.
* Move RecordingContext and asMap from TestHelper into
  DatabaseAdaptorTest and MetadataColumnsTest, respectively.
