+++ test/com/google/enterprise/adaptor/database/ResponseGeneratorTest.java
@@ -37,7 +37,8 @@ import static org.junit.Assume.assumeFalse;
 import static org.junit.Assume.assumeTrue;
 
 import com.google.enterprise.adaptor.InvalidConfigurationException;
-import com.google.enterprise.adaptor.database.RecordingResponse.State;
+import com.google.enterprise.adaptor.testing.RecordingResponse;
+import com.google.enterprise.adaptor.testing.RecordingResponse.State;
 
 import org.junit.After;
 import org.junit.Rule;
